+++ lib/vyasa/adapters/binding.ex
@@ -6,16 +6,21 @@ defmodule Vyasa.Adapters.Binding do
   use Ecto.Schema
   import Ecto.Changeset
 
-  alias Vyasa.Written.{Verse, Source, Chapter}
+  alias Vyasa.Written.{Source, Chapter, Verse, Translation}
+  alias Vyasa.Sangh.{Comment}
 
-
-  embedded_schema do
+  @primary_key {:id, Ecto.UUID, autogenerate: true}
+  schema "bindings" do
+    field :w_type, Ecto.Enum, values: [:quote, :timestamp, :null]
+    field :field_key, :string
     belongs_to :verse, Verse, foreign_key: :verse_id, type: :binary_id
     belongs_to :chapter, Chapter, type: :integer, references: :no, foreign_key: :chapter_no
     belongs_to :source, Source, foreign_key: :source_id, type: :binary_id
+    belongs_to :translation, Translation, foreign_key: :translation_id, type: :binary_id
+    belongs_to :comment, Comment, foreign_key: :comment_id, type: :binary_id
+    belongs_to :comment_bind, Comment, foreign_key: :comment_bind_id, type: :binary_id
 
     embeds_one :window, Window, on_replace: :delete do
-      field(:key, :string) # from target assoc hierarchy source >> chapter >> verse >> translations tie to matrix
       field(:line_number, :integer)
       field(:start, :integer)
       field(:end, :integer)
@@ -30,6 +35,50 @@ defmodule Vyasa.Adapters.Binding do
     event
     |> cast(attrs, [:verse_id, :voice_id, :source_id])
   end
+
+
+  def bind_comment_changeset(%__MODULE__{} = binding, attrs) do
+    binding
+    |> cast(attrs, [:w_type, :verse_id, :chapter_no, :source_id])
+    |> typed_window_switch(attrs)
+    |> Map.put(:repo_opts, [on_conflict: {:replace_all_except, [:id]}, conflict_target: :id])
+  end
+
+
+  #when type changes
+  def typed_window_switch(changeset, %{w_type: type}), do: typed_window_switch(changeset, %{"w_type" => type})
+  def typed_window_switch(changeset, %{"w_type" => type}) do
+    window_changeset = case type do
+                            "quote" ->
+                              &quote_changeset(&1, &2)
+                            "timestamp" ->
+                              &timestamp_changeset(&1, &2)
+                            _ ->
+                              &null_changeset(&1, &2)
+                          end
+
+    cast_embed(changeset, :window, with: window_changeset)
+  end
+
+  def typed_window_switch(changeset, _attrs), do: validate_required(changeset, [:w_type])
+
+  def quote_changeset(structure, attrs) do
+    structure
+    |> cast(attrs, [:line_number, :start, :end, :quote])
+    |> validate_required([:line_number, :quote])
+  end
+
+  def timestamp_changeset(structure, attrs) do
+    structure
+    |> cast(attrs, [:start_time, :end_time])
+    |> validate_required([:start_time, :end_time])
+  end
+
+  def null_changeset(structure, attrs) do
+    structure
+    |> cast(attrs, [])
+  end
+
 
   def cast(attrs) do
     %__MODULE__{}
